added license error as general error & ua merge
